Updates + Implementing Unit Testing
I started implementing JUnit testing for the base code. While it would
have been a much better approach to have written them as I went along, I
was unfortunately on time constraints for the original Udacity project.
Now that I no longer have that time constraint, I want to take the time
to implement project wide testing, which includes Connected and JUnit
testing. This commit has a handful of the new tests I'm writing for the
application. Future commits may contain refactored code to create more
testable code.

There is now more control over the Drive API as well. If the app is
unable to connect to the Drive, it throws up a dialog that prompt the
user to cancel or try again.

Lastly, the font for the editor text has been changed to monospace to
make it easier to read. Besides, assembly looks pretty awesome using
monospace.